Added replace of blocks and link connections
#Bug 2 links in out port than replace and its bugged

+++ src/Graphics/LineMesh.java
@@ -12,7 +12,7 @@ public class LineMesh
 
     private void CalcRects(Point2D startPoint, Point2D finalPoint)
     {
-        int middleX = (startPoint.X + finalPoint.X) /2;
+        double middleX = (startPoint.X + finalPoint.X) /2;
         _first = new Rect(startPoint.X,startPoint.Y+1,middleX,3);
         _middle = new Rect(middleX-1,startPoint.Y,3,finalPoint.Y);
         _last = new Rect(middleX,finalPoint.Y+1,middleX,3);
